Created film seeds with dumby image attachments

+++ db/migrate/20201120205001_create_films.rb
@@ -3,7 +3,7 @@ class CreateFilms < ActiveRecord::Migration[5.2]
     create_table :films do |t|
       t.string :title, uniqueness: true
       t.string :synopsis, null: false
-      t.integer :rating, null: false, default: 0
+      t.decimal :rating, null: false, default: 0
       t.string :Language, null: false
       t.date :Release_date, null: false
